btrfs-util: move btrfs_defrag_fd() from fd-util.[ch]
After d71ece3f0b85c7a3decc50143b68ac07fc5831ae, the function is not used
in libbasic or libsystemd anymore. Let's move it to more appropriate
place.
